test(plugins/plugin-k8s): remove 'wipe resource' and parallelize k status test
Fixes #1457
