+++ plugins/plugin-k8s/src/test/k8s/get-pods-with-watch.ts
@@ -16,7 +16,6 @@
 
 import * as common from '@kui-shell/core/tests/lib/common'
 import { cli, expectSubset, selectors, sidecar } from '@kui-shell/core/tests/lib/ui'
-import { wipe, waitTillNone } from '@kui-shell/plugin-k8s/tests/lib/k8s/wipe'
 import { createNS, allocateNS, deleteNS } from '@kui-shell/plugin-k8s/tests/lib/k8s/utils'
 
 /** name of the pod */
