+++ plugins/plugin-k8s/src/test/k8s1/kedit.ts
@@ -16,7 +16,6 @@
 
 import * as common from '@kui-shell/core/tests/lib/common'
 import { cli, selectors, sidecar, expectYAML, getValueFromMonaco } from '@kui-shell/core/tests/lib/ui'
-import { wipe, waitTillNone } from '@kui-shell/plugin-k8s/tests/lib/k8s/wipe'
 
 import { Application } from 'spectron'
