+++ plugins/plugin-k8s/src/test/k8s1/status.ts
@@ -16,27 +16,29 @@
 
 import * as common from '@kui-shell/core/tests/lib/common'
 import { cli, selectors } from '@kui-shell/core/tests/lib/ui'
-import { wipe, waitTillNone } from '@kui-shell/plugin-k8s/tests/lib/k8s/wipe'
+import { createNS, waitTillNone } from '@kui-shell/plugin-k8s/tests/lib/k8s/utils'
 import { kubectl, cli as kui, CLI } from '@kui-shell/core/tests/lib/headless'
 
 import { dirname } from 'path'
 const ROOT = dirname(require.resolve('@kui-shell/plugin-k8s/tests/package.json'))
 
-// NOTE: k status only works in defualt namespace
 const doHeadless = (ctx: common.ISuite, impl: CLI) => {
   before(common.before(ctx, { noApp: true }))
+  const ns = createNS()
 
-  it('should wipe k8s', () => {
-    return wipe(ctx, impl)
+  it(`should create namespace ${ns} via kubectl`, () => {
+    return impl.do(`kubectl create namespace ${ns}`, ctx.app)
+      .then(impl.expectOK(ns))
+      .catch(common.oops(ctx))
   })
 
   it('should create sample pod from local file', () => {
-    return impl.do(`kubectl create -f ${ROOT}/data/k8s/headless/pod.yaml`, ctx.app)
+    return impl.do(`kubectl create -f ${ROOT}/data/k8s/headless/pod.yaml -n ${ns}`, ctx.app)
       .then(impl.expectOK('nginx'))
       .catch(common.oops(ctx))
   })
 
-  const alternatives = ['k status', 'kubectl status', 'k list', 'kubectl list']
+  const alternatives = [`k status pods -n ${ns}`, `kubectl status pods -n ${ns}`, `k list pods -n ${ns}`, `kubectl list pods -n ${ns}`]
   alternatives.forEach(cmd => {
     it(`should list the new pod via the "${cmd}"`, () => {
       return impl.do(cmd, ctx.app)
@@ -46,17 +48,25 @@ const doHeadless = (ctx: common.ISuite, impl: CLI) => {
   })
 
   it('should delete the new pod by yaml', () => {
-    return impl.do(`kubectl delete -f ${ROOT}/data/k8s/headless/pod.yaml`, ctx.app)
+    return impl.do(`kubectl delete -f ${ROOT}/data/k8s/headless/pod.yaml  -n ${ns}`, ctx.app)
       .then(impl.expectOK('pod "nginx" deleted'))
-      .then(waitTillNone('pods', impl))
+      .then(() => waitTillNone('pods', impl)(ctx.app))
       .catch(common.oops(ctx))
   })
 
   it('should NOT list the new pod via the status command', () => {
-    return impl.do('k status', ctx.app)
+    return impl.do(`k status pods -n ${ns}`, ctx.app)
       .then(impl.expectJustOK())
       .catch(common.oops(ctx))
   })
+
+  if (!process.env.TRAVIS_JOB_ID) {
+    it(`should delete the namespace ${ns} `, () => {
+      return impl.do(`kubectl delete namespace ${ns}`, ctx.app)
+        .then(impl.expectOK(`namespace "${ns}" deleted`))
+        .catch(common.oops(ctx))
+    })
+  }
 }
 
 describe('k status kubectl kui headless mode', function (this: common.ISuite) {
