+++ plugins/plugin-k8s/src/test/k8s2/get-all-namespaces.ts
@@ -16,8 +16,7 @@
 
 import * as common from '@kui-shell/core/tests/lib/common'
 import { cli, selectors, sidecar } from '@kui-shell/core/tests/lib/ui'
-import { waitTillNone } from '@kui-shell/plugin-k8s/tests/lib/k8s/wipe'
-import { defaultModeForGet, createNS as create } from '@kui-shell/plugin-k8s/tests/lib/k8s/utils'
+import { defaultModeForGet, createNS as create, waitTillNone } from '@kui-shell/plugin-k8s/tests/lib/k8s/utils'
 
 const synonyms = ['kubectl']
