+++ plugins/plugin-k8s/src/test/k8s2/headless-create-pod.ts
@@ -15,9 +15,8 @@
  */
 
 import * as common from '@kui-shell/core/tests/lib/common'
-import { waitTillNone } from '@kui-shell/plugin-k8s/tests/lib/k8s/wipe'
 import { kubectl, cli as kui, CLI } from '@kui-shell/core/tests/lib/headless'
-import { createNS } from '@kui-shell/plugin-k8s/tests/lib/k8s/utils'
+import { createNS, waitTillNone } from '@kui-shell/plugin-k8s/tests/lib/k8s/utils'
 
 import { dirname } from 'path'
 const ROOT = dirname(require.resolve('@kui-shell/plugin-k8s/tests/package.json'))
@@ -69,7 +68,7 @@ const doHeadless = (ctx: common.ISuite, impl: CLI) => {
     it(`should delete the new pod by yaml via ${kubectl}`, () => {
       return impl.do(`${kubectl} delete -f ${ROOT}/data/k8s/headless/pod.yaml ${inNamespace}`, ctx.app)
         .then(impl.expectOK('pod "nginx" deleted'))
-        .then(waitTillNone('pods', impl, undefined, undefined, inNamespace))
+        .then(() => waitTillNone('pods', impl, undefined, undefined, inNamespace)(ctx.app))
         .catch(common.oops(ctx))
     })
 
@@ -82,7 +81,7 @@ const doHeadless = (ctx: common.ISuite, impl: CLI) => {
     it(`should delete the new pod by name via ${kubectl}`, () => {
       return impl.do(`${kubectl} delete pod nginx ${inNamespace}`, ctx.app)
         .then(impl.expectOK('pod "nginx" deleted'))
-        .then(waitTillNone('pods', impl, undefined, undefined, inNamespace))
+        .then(() => waitTillNone('pods', impl, undefined, undefined, inNamespace)(ctx.app))
         .catch(common.oops(ctx))
     })
