+++ plugins/plugin-openwhisk/src/test/openwhisk3/popup.ts
@@ -16,7 +16,6 @@
 
 import * as common from '@kui-shell/core/tests/lib/common'
 import { cli, selectors, sidecar } from '@kui-shell/core/tests/lib/ui'
-import { wipe, waitTillNone } from '@kui-shell/plugin-k8s/tests/lib/k8s/wipe'
 import * as openwhisk from '@kui-shell/plugin-openwhisk/tests/lib/openwhisk/openwhisk'
 const { localDescribe } = common
 import { readFileSync } from 'fs'
